Logging update

+++ onewire_logger_db_fb.py
@@ -7,9 +7,18 @@ import firebase_admin_file
 from os import chdir as os_chdir, path as os_path
 os_chdir(os_path.dirname(os_path.abspath(__file__)))
 
+# Enable/disable rebooting when sensors aren't being detected.
+REBOOT_ON_SENSOR_COUNT = False
+
+# Enable/disable debug notifications
+DEBUG_NOTIFICATION = False
+
+# Enable/disable debugging print statements
+DEBUG_PRINT = False
+
 SENSOR_NAMES = ReadingObj.SENSOR_NAMES
 # Sleep interval between readings
-#  750ms conversion time + manual 250ms sleep = 1s per sensor read time.
+#  750ms conversion time + manual 250ms sleep = around 1s per sensor read time.
 # sensor accuracy = +- 0.0625
 INTERVAL = 18.9
 
@@ -17,7 +26,7 @@ INTERVAL = 18.9
 ROUNDING = 2
 
 # Number of physically connected sensors
-SENSOR_COUNT = 10
+SENSOR_COUNT = 9
 
 # Maximum errors before system reboot.
 MAX_ERRORS = 10
@@ -46,7 +55,7 @@ error_count_other = 0
 # (10 DEBUG, 20 INFO, 30 WARNING, 40 ERROR, 50 CRITICAL)
 logging_basicConfig(
     level=const_ERROR,
-    format='\n%(asctime)s - %(levelname)s - %(message)s',
+    format='%(asctime)s - %(levelname)s - %(message)s',
     datefmt='%Y-%m-%d %H:%M:%S',
     filename='/home/luke/Desktop/Script/Logs/error_logging.txt',
     filemode='a'
@@ -82,29 +91,50 @@ while True:
         # create time string
         date_time_now = strftime('%Y-%m-%d %H:%M:%S')
 
-        # Reboot if error count reaches set value. One or more sensors are not detected.
-        # Rebooting helps when sensors are not being detected. Log the rest of the sensors.
+        '''
+        Reboot if error count reaches set value. One or more sensors are not detected.
+        Rebooting helps when sensors are not being detected. Log the rest of the sensors.
+        Most of the time its the same sensor that isn't detected. Long cable issue. We can ignore this one.
+        '''
         if num_of_sensors != SENSOR_COUNT:
-            #print(f"____ERROR____   only found {num_of_sensors} / {SENSOR_COUNT} sensors")
+            if DEBUG_PRINT:
+                print(f"__ERROR  only found {num_of_sensors} / {SENSOR_COUNT} sensors")
+            
+            
             error_count_sensors += 1
             if error_count_sensors >= MAX_ERRORS:
-                logging_info(f"_ERROR_ REBOOTING because detected only {num_of_sensors} / {SENSOR_COUNT} sensors.")
-                firebase_admin_file.send_notification('debug', 'error', f"REBOOTING because detected only {num_of_sensors} / {SENSOR_COUNT} sensors.")
-                sleep(30)
-                subprocess_call('sudo reboot', shell=True)
+                logging_error(f'_ERROR_  detected only {num_of_sensors} / {SENSOR_COUNT} sensors.')
+                # reset 
+                error_count_sensors = 0
+                
+                if DEBUG_NOTIFICATION:
+                    firebase_admin_file.send_notification('debug', 'error', \
+                    f'REBOOTING @ {datetime.now().strftime("%a %I:%M %p")} \
+                    Detected only {num_of_sensors} / {SENSOR_COUNT} sensors.')
+                if REBOOT_ON_SENSOR_COUNT:
+                    logging_error(f'_ERROR_ REBOOTING because detected only {num_of_sensors} / {SENSOR_COUNT} sensors.')
+
+                    sleep(30)
+                    subprocess_call('sudo reboot', shell=True)
             
-            else:
-                logging_info(f"_ERROR_ Detected only {num_of_sensors} / {SENSOR_COUNT} sensors.  \
-                error Count {error_count_sensors}") 
+            elif DEBUG_NOTIFICATION and num_of_sensors < (SENSOR_COUNT - 1) and error_count_sensors >= MAX_ERRORS:
+                # Notify if more than one sensor isn't being detected.
+                firebase_admin_file.send_notification('debug', 'Warning', \
+                f'Detected only {num_of_sensors} / {SENSOR_COUNT} sensors.')
+                error_count_sensors = 0
+                
+        
         '''
         Begin reading sensors
         '''
         # Read Available Sensors
-        #print("reading sensors...")
+        if DEBUG_PRINT:
+            print("__reading sensors...")
         readingObj = read_sensors(sensor_obj, previousReadingObj, date_time_now, num_of_sensors, ROUNDING)
         sensor_vals_tuple = readingObj.get_solar_tuple()
         sensor_vals_string = readingObj.get_solar_str()
-        
+        if DEBUG_PRINT:
+            print(sensor_vals_string)
         '''
         -Stage 1 recovery-
         Keep a previous reading in memory incase of sensor read failure.
@@ -132,9 +162,15 @@ while True:
 
         # Storage Location 4 - Firebase firestore
         '''
-        Todo: deal with connection exceptions. record timestamps where exception occurs. when connection is restored, write all failed lines to db.
+        TODO: deal with connection exceptions. record timestamps where exception occurs.
+        when connection is restored, write all failed lines to db.
         Merging with previous hours will need to be done.
+        
+        if sensor isn't detected, the default value of None will be written to databases.
+        This crashes the android app. Either update app or do not write Nones.
+        -8_25_2024 workaround is to use 0.01 instead of None as default. not ideal but prevents app from crashing.
         '''
+        
         lastHourDocumentRef = firebase_admin_file.write_line(date_time_utc, sensor_vals_string, lastHourDocumentRef, readingObj)
         
     except ValueError as error:
@@ -145,21 +181,25 @@ while True:
          Write lines to file and upload them when connection is restored.
          Check if file exists. If so, read line by line and firebase.write_line
         '''
-        logging_error(f"An exception of type {type(error).__name__} occurred: {str(error)}\n    error_count: {error_count_other}\n\n", exc_info=True)
+        logging_error(f"An exception of type {type(error).__name__} occurred: {str(error)}\n    error_count: {error_count_other}\n\n", exc_info=False)
         log_reading(sensor_vals_string)
         error_count_other += 1
 
     except Exception as error:
-        print(f"---------ERROR--------{type(error)}--------\n {error}")
+        if DEBUG_PRINT:
+            print(f"---------ERROR--------{type(error)}--------\n {error}")
         error_count_other += 1        
-        logging_error(f"An exception of type {type(error).__name__} occurred: {str(error)}\n    error_count: {error_count_other}\n\n", exc_info=True)     
+        logging_error(f"An exception of type {type(error).__name__} occurred: {str(error)}\n    error_count: {error_count_other}\n\n", exc_info=False)     
         
-    finally:        
-        #print ('___waiting {} seconds...'.format(INTERVAL))
+    finally:
+        if DEBUG_PRINT:
+            print ('__waiting {} seconds...'.format(INTERVAL))
         if error_count_other >= MAX_ERRORS:
             logging_error(f"REBOOTING due to error count of {error_count_other}.")
             try:
-                firebase_admin_file.send_notification('debug', 'error', f"REBOOTING due to error count of {error_count_other}.")
+                # try/except here because we cant have errors in the finally block.
+                if DEBUG_NOTIFICATION:
+                    firebase_admin_file.send_notification('debug', 'error', f'REBOOTING @ {datetime.now().strftime("%a %I:%M %p")} due to error count of {error_count_other}.')
             except:
                 logging_error(f"Failed to send notification")
                 pass
